[#23] Refactor: 컴포넌트 구조 개선
- Root 페이지를 만들어 Header를 포함하는 기본적인 페이지 레이아웃을 만들었습니다
- 메인 페이지를 크게 Title, FilterBox, TransactionTable 3개의 컴포넌트로 분리하였습니다
- 한 파일에 모두 담기 힘든  컴포넌트들은 같은  이름으로 폴더를 만들어 컴포넌트를 쪼개고 폴더에 담았습니다
- Search 컴포넌트 위치를 FilterBox쪽으로 옮겼습니다
