+++ src/components/FilterBox/index.tsx
@@ -1,11 +1,14 @@
-import { Stack } from '@mui/material';
+import { Stack, Box } from '@mui/material';
 
 import DateFilter from './DateFilter';
+import SearchFilter from './SearchFilter';
 import StatusFilter from './StatusFilter';
 
 const FilterBox = () => {
   return (
     <Stack direction='row' justifyContent='space-between'>
+      <SearchFilter />
+      <Box flex={1} />
       <DateFilter />
       <StatusFilter />
     </Stack>
