+++ src/components/Header.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Header = () => {
+  return <div>Header</div>;
+};
+
+export default Header;
