+++ src/components/TableTitle.tsx
@@ -1,0 +1,13 @@
+import { Typography, Box } from '@mui/material';
+
+const TableTitle = () => {
+  return (
+    <Box marginBottom={4}>
+      <Typography variant='h5' fontWeight='bold'>
+        Transaction History
+      </Typography>
+    </Box>
+  );
+};
+
+export default TableTitle;
