+++ src/components/TransactionTable/StyledTableCell.tsx
@@ -1,0 +1,15 @@
+import { TableCell, tableCellClasses, styled } from '@mui/material';
+
+const StyledTableCell = styled(TableCell)(() => ({
+  [`&.${tableCellClasses.head}`]: {
+    borderBottom: '1px solid #cecece',
+    color: '#1B2E57',
+    fontWeight: 'bold',
+  },
+  [`&.${tableCellClasses.body}`]: {
+    borderBottom: '1px solid #eeefef',
+    color: '#1B2E57',
+  },
+}));
+
+export default StyledTableCell;
