+++ src/pages/Root.tsx
@@ -1,0 +1,15 @@
+import { Container } from '@mui/material';
+import { Outlet } from 'react-router-dom';
+
+import Header from '@/components/Header';
+
+const Root = () => {
+  return (
+    <Container>
+      <Header />
+      <Outlet />
+    </Container>
+  );
+};
+
+export default Root;
