+++ src/routes/index.tsx
@@ -4,8 +4,22 @@ import loader from './loader';
 
 import Error from '@/pages/Error';
 import Main from '@/pages/Main';
+import Root from '@/pages/Root';
 
-const routes: RouteObject[] = [{ path: '/', element: <Main />, errorElement: <Error />, loader }];
+const routes: RouteObject[] = [
+  {
+    path: '/',
+    element: <Root />,
+    errorElement: <Error />,
+    children: [
+      {
+        loader,
+        index: true,
+        element: <Main />,
+      },
+    ],
+  },
+];
 
 const router = createBrowserRouter(routes);
